Added function for manage product in the page admin, made the account etc

+++ dz_8/engine/reviews.php
@@ -4,30 +4,24 @@ require_once LIB_DIR . 'db.php';
 require_once LIB_DIR . 'functions.php';
 
 function manageReview($values = []) {
+	if(!empty($values)) {
+		//var_dump($values);
 	foreach($values as $key => $val) {
 		$valuesNew[$key] = checkValue($val);
 	}
-	if($values['addreview'] !== null) {
-		//echo $date;
+	if(isset($values['addreview']) && $values['addreview'] !== null) {
 		$sql = "insert into review (`name`, `text`, `created_at`) values ('" . $valuesNew['name'] . "', '" . $valuesNew['review'] . "', '" . $valuesNew['date'] . "');";
 		//echo $sql;
 	}
-	if($values['delreview'] !== null) {
-		/*echo "<pre>";
-		var_dump($valuesNew);
-		echo "</pre>";*/
+	if(isset($values['delreview']) && $values['delreview'] !== null) {
 		$sql = "delete from review where id = " . $valuesNew['id'];
-		//echo $sql;
 	}
 	executeQuery($sql, DB_REVIEW);
+	}
 }
 function showReviews() {
 	$sql = "select * from review order by `created_at` desc;";
 	$res = getAssocResult($sql, DB_REVIEW);
-	/*echo "<pre>";
-	var_dump($res);
-	echo "</pre>";
-	echo "<br>" . $sql;*/
 	return $res;
 	
 }
